+++ src/cqterrain/tile/__init__.py
@@ -1,3 +1,4 @@
+from .apricorn import apricorn
 from .octagonWithDots import octagon_with_dots
 from .octagonWithDots import octagon_with_dots_2
 from .basketweave import basketweave