+++ src/cqterrain/tile/apricorn.py
@@ -1,0 +1,49 @@
+import cadquery as cq
+
+def _make_lines(
+    length, 
+    height,
+    line_width,
+    center_radius
+):
+    line = cq.Workplane("XY").box(length, line_width, height)
+    center = cq.Workplane("XY").cylinder(height, center_radius+line_width)
+    cut_center = cq.Workplane("XY").cylinder(height, center_radius)
+    return center.union(line).cut(cut_center)
+
+def apricorn(
+    length = 30, 
+    width = 25, 
+    height = 4,
+    line_width = 2,
+    line_depth = .5,
+    center_radius = None,
+    width_radius_divisor = 4
+):
+    if not center_radius:
+        center_radius = width / width_radius_divisor
+    
+    lines = _make_lines(
+        length, 
+        height,
+        line_width,
+        center_radius
+    )
+    
+    cut_lines = _make_lines(
+        length - line_depth*2,
+        height - line_depth,
+        line_width,
+        center_radius
+    ).translate((0,0,-1*(line_depth/2)))
+    
+    combined_lines = lines.cut(cut_lines)
+
+    outline = cq.Workplane("XY").box(length, width, height)
+    
+    apricorn_tile = (
+        cq.Workplane("XY")
+        .union(outline)
+        .cut(combined_lines)
+    )
+    return apricorn_tile