+++ changes.md
@@ -2,11 +2,12 @@
 
 ## 0.3.1
 * Added pipe.corrugated_straight
+* Added tile.apricorn
 * Added tile.bolt_panel
 * Added tile.conduit
 * Added tile.carton
 * Added tile.carton2
-* Tile documentation and examples for bolt_panel, glyph, conduit, carton, carton2
+* Tile documentation and examples for apricorn, bolt_panel, glyph, conduit, carton, carton2
 
 ## 0.3.0
 * Upped cadqueryhelper version 0.2.0