+++ changes.md
@@ -4,10 +4,11 @@
 * Added pipe.corrugated_straight
 * Added tile.apricorn
 * Added tile.bolt_panel
+* Added tile.charge
 * Added tile.conduit
 * Added tile.carton
 * Added tile.carton2
-* Tile documentation and examples for apricorn, bolt_panel, glyph, conduit, carton, carton2
+* Tile documentation and examples for apricorn, bolt_panel, charge, glyph, conduit, carton, carton2
 
 ## 0.3.0
 * Upped cadqueryhelper version 0.2.0